New pool, weapon stats, and some folder organization
The pool of objects require a gameobject either inheriting from pool or having the pool component.

also the Pool manager is in the resources folder, it needs the reference for the pool object and a name so the pool knows where to storage itself.

The pool manager has a method named Item_Call that calls the type of item the name points at, the Item_Dismiss method  is to get rid of the gameobject in question to return it to the pool

+++ Assets/Scripts/Systems/Health Systems/HealthSystem.cs
@@ -5,11 +5,11 @@ using UnityEngine.Events;
 
 public class HealthSystem : MonoBehaviour
 {
-    public int maxHealthPoints;
-    public int healthPoints;
+    public float maxHealthPoints;
+    public float healthPoints;
     public bool isDead;
 
-    public UnityEvent<int> OnHealthChanged;
+    public UnityEvent<float> OnHealthChanged;
 
     void Start()
     {
@@ -17,7 +17,7 @@ public class HealthSystem : MonoBehaviour
 
     }
 
-    public void Damage(int damageCaused)
+    public void Damage(float damageCaused)
     {
         healthPoints -= damageCaused;
         OnHealthChanged.Invoke(healthPoints);
